solve finish problem issue
